# perf: DirtyObjects (V2)