--- a/Assets/Mirror/Core/NetworkServer.cs
+++ b/Assets/Mirror/Core/NetworkServer.cs
@@ -1,8 +1,10 @@
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Linq;
 using Mirror.RemoteCalls;
 using UnityEngine;
+using Debug = UnityEngine.Debug;
 
 namespace Mirror
 {
@@ -1810,6 +1812,7 @@ namespace Mirror
         internal static readonly List<NetworkConnectionToClient> connectionsCopy =
             new List<NetworkConnectionToClient>();
 
+        static Stopwatch watch = new Stopwatch();
         static void Broadcast()
         {
             // copy all connections into a helper collection so that
@@ -1823,11 +1826,15 @@ namespace Mirror
             connectionsCopy.Clear();
             connections.Values.CopyTo(connectionsCopy);
 
+            watch.Restart();
+
             // broadcast spawned entities
             BroadcastDirtySpawned();
 
             // flush all connection's batched messages
             FlushConnections();
+
+            Debug.Log($"Broadcast took {watch.Elapsed.TotalMilliseconds:F1} ms");
 
             // TODO this is way too slow because we iterate ALL spawned :/
             // TODO this is way too complicated :/