new implementation with:
- keeps SyncInterval
- DirtyObjects as List for more memory friendly layout
- faster

results still inconclusive:

```
IL2CPP build, one client, syncInterval=0.1, server camera disabled(!), telepathy

  100k objects,    _0% sleeping:_
    BEFORE:   4.4 - 27.9 ms broadcast | 33 FPS
    DIRTYOBJ: 2.6 - 10.2 ms broadcast | 35 FPS

  100k objects,   _50% sleeping:_
    BEFORE:   4.7 - 18.5 ms broadcast | 37 FPS
    DIRTYOBJ: 2.0 -  5.9 ms broadcast | 39 FPS

  100k objects,   _80% sleeping:_
    BEFORE:   4.6 - 10.2 ms broadcast | 40 FPS
    DIRTYOBJ: 1.8 -  3.5 ms broadcast | 41 FPS

  100k objects,  _100% sleeping:_
    BEFORE:   4.5 -  5.0 ms broadcast | 46 FPS
    DIRTYOBJ:        0.0 ms broadcast | 60 FPS
```